# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * La Liga Brasil API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type ArticleStatus = (typeof ArticleStatus)[keyof typeof ArticleStatus];

export const ArticleStatus = {
  draft: "draft",
  published: "published",
  scheduled: "scheduled",
} as const;

export interface Article {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  coverImage?: string | null;
  status: ArticleStatus;
  featured: boolean;
  breakingNews: boolean;
  category: string;
  authorName: string;
  teamId?: number | null;
  teamName?: string | null;
  teamSlug?: string | null;
  sourceUrl?: string | null;
  sourceName?: string | null;
  publishedAt?: string | null;
  scheduledAt?: string | null;
  createdAt: string;
  updatedAt: string;
  viewCount: number;
}

export interface ArticleListResponse {
  articles: Article[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface Team {
  id: number;
  name: string;
  slug: string;
  shortName: string;
  city: string;
  stadium: string;
  foundedYear?: number | null;
  description?: string | null;
  logoUrl?: string | null;
  primaryColor: string;
  secondaryColor: string;
  articleCount: number;
}

export type TeamWithArticles = Team & {
  recentArticles: Article[];
};

export interface Category {
  id: number;
  name: string;
  slug: string;
  articleCount: number;
}

export type CreateArticleRequestStatus =
  (typeof CreateArticleRequestStatus)[keyof typeof CreateArticleRequestStatus];

export const CreateArticleRequestStatus = {
  draft: "draft",
  published: "published",
  scheduled: "scheduled",
} as const;

export interface CreateArticleRequest {
  title: string;
  excerpt: string;
  content: string;
  coverImage?: string | null;
  status: CreateArticleRequestStatus;
  featured: boolean;
  breakingNews: boolean;
  category: string;
  authorName: string;
  teamId?: number | null;
  sourceUrl?: string | null;
  sourceName?: string | null;
  scheduledAt?: string | null;
}

export interface ScheduleArticleRequest {
  scheduledAt: string;
}

export interface CreateTeamRequest {
  name: string;
  slug: string;
  shortName: string;
  city: string;
  stadium: string;
  foundedYear?: number | null;
  description?: string | null;
  logoUrl?: string | null;
  primaryColor: string;
  secondaryColor: string;
}

export interface NewsSource {
  id: string;
  name: string;
  url: string;
  language: string;
  description: string;
  active: boolean;
}

export interface FetchSourceRequest {
  sourceId: string;
  maxArticles?: number;
  autoImport?: boolean;
}

export interface ScrapedArticle {
  title: string;
  originalTitle: string;
  excerpt: string;
  content: string;
  originalUrl: string;
  sourceName: string;
  coverImage?: string | null;
  publishedAt?: string | null;
  imported: boolean;
}

export interface FetchSourceResponse {
  articles: ScrapedArticle[];
  imported: number;
  total: number;
}

export interface TranslateRequest {
  title: string;
  content: string;
  excerpt: string;
  sourceName: string;
  sourceLanguage?: string;
}

export interface TranslateResponse {
  title: string;
  content: string;
  excerpt: string;
}

export interface SiteStats {
  totalArticles: number;
  publishedArticles: number;
  draftArticles: number;
  scheduledArticles: number;
  totalTeams: number;
  totalViews: number;
  recentArticles: Article[];
}

export type ListArticlesParams = {
  page?: number;
  limit?: number;
  category?: string;
  teamId?: number;
  featured?: boolean;
  search?: string;
};

export type AdminListArticlesParams = {
  page?: number;
  limit?: number;
  status?: AdminListArticlesStatus;
  category?: string;
  teamId?: number;
  source?: string;
  search?: string;
};

export type AdminListArticlesStatus =
  (typeof AdminListArticlesStatus)[keyof typeof AdminListArticlesStatus];

export const AdminListArticlesStatus = {
  draft: "draft",
  published: "published",
  scheduled: "scheduled",
} as const;