# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: La Liga Brasil API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List published articles
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
        - name: teamId
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleListResponse"
  /articles/{slug}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get article by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /admin/articles:
    get:
      operationId: adminListArticles
      tags: [admin]
      summary: List all articles (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, scheduled]
        - name: category
          in: query
          schema:
            type: string
        - name: teamId
          in: query
          schema:
            type: integer
        - name: source
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleListResponse"
    post:
      operationId: createArticle
      tags: [admin]
      summary: Create article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleRequest"
      responses:
        "201":
          description: Created article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /admin/articles/{id}:
    get:
      operationId: adminGetArticle
      tags: [admin]
      summary: Get article by id (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    put:
      operationId: updateArticle
      tags: [admin]
      summary: Update article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleRequest"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    delete:
      operationId: deleteArticle
      tags: [admin]
      summary: Delete article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/articles/{id}/publish:
    post:
      operationId: publishArticle
      tags: [admin]
      summary: Publish article immediately
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /admin/articles/{id}/schedule:
    post:
      operationId: scheduleArticle
      tags: [admin]
      summary: Schedule article for future publish
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleArticleRequest"
      responses:
        "200":
          description: Scheduled article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Article:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        coverImage:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, scheduled]
        featured:
          type: boolean
        breakingNews:
          type: boolean
        category:
          type: string
        authorName:
          type: string
        teamId:
          type: integer
          nullable: true
        teamName:
          type: string
          nullable: true
        teamSlug:
          type: string
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        sourceName:
          type: string
          nullable: true
        publishedAt:
          type: string
          nullable: true
        scheduledAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        viewCount:
          type: integer
      required:
        - id
        - title
        - slug
        - excerpt
        - content
        - status
        - featured
        - breakingNews
        - category
        - authorName
        - createdAt
        - updatedAt
        - viewCount
    ArticleListResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - articles
        - total
        - page
        - limit
        - totalPages
...
    CreateArticleRequest:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        coverImage:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, scheduled]
        featured:
          type: boolean
        breakingNews:
          type: boolean
        category:
          type: string
        authorName:
          type: string
        teamId:
          type: integer
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        sourceName:
          type: string
          nullable: true
        scheduledAt:
          type: string
          nullable: true
      required:
        - title
        - excerpt
        - content
        - status
        - featured
        - breakingNews
        - category
        - authorName
    ScheduleArticleRequest:
      type: object
      properties:
        scheduledAt:
          type: string
      required:
        - scheduledAt
...