# lib/api-zod/src/generated/types/createArticleRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * La Liga Brasil API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateArticleRequestStatus } from "./createArticleRequestStatus";

export interface CreateArticleRequest {
  title: string;
  excerpt: string;
  content: string;
  coverImage?: string | null;
  status: CreateArticleRequestStatus;
  featured: boolean;
  breakingNews: boolean;
  category: string;
  authorName: string;
  teamId?: number | null;
  sourceUrl?: string | null;
  sourceName?: string | null;
  scheduledAt?: string | null;
}